updated swc compilation target
