+++ index.html
@@ -16,6 +16,7 @@
 
   <body>
     <div id="root" />
+    <div id="target-canvas"></div>
     <script type="module" src="/lib/es6/src/index.bs.js"></script>
   </body>
 </html>
