+++ src/pages/Home.res
@@ -1,6 +1,9 @@
 open Layout
+open ThreeCanvas
 
 @react.component
 let make = () => {
-  <Layout> <div> {"Testing Three on Homepage"->React.string} </div> </Layout>
+  <Layout> <div> {"Testing Three on Homepage"->React.string} </div>
+  <ThreeCanvas />
+  </Layout>
 }
